AM156

Approved-by: Christina Giampalmo <cgiampalmo@glaciersecurity.com>

+++ src/main/java/com/glaciersecurity/glaciermessenger/services/XmppConnectionService.java
@@ -302,7 +302,7 @@ public class XmppConnectionService extends Service {
 			fetchBookmarks(account);
 			final boolean flexible = account.getXmppConnection().getFeatures().flexibleOfflineMessageRetrieval();
 			final boolean catchup = getMessageArchiveService().inCatchup(account);
-			if (flexible && catchup) {
+			if (flexible && catchup && account.getXmppConnection().isMamPreferenceAlways()) {
 				sendIqPacket(account, mIqGenerator.purgeOfflineMessages(), (acc, packet) -> {
 					if (packet.getType() == IqPacket.TYPE.RESULT) {
 						Log.d(Config.LOGTAG, acc.getJid().asBareJid() + ": successfully purged offline messages");
@@ -872,15 +872,19 @@ public class XmppConnectionService extends Service {
 	@SuppressLint("NewApi")
 	@SuppressWarnings("deprecation")
 	public boolean isInteractive() {
+        try {
 		final PowerManager pm = (PowerManager) getSystemService(Context.POWER_SERVICE);
 
-		final boolean isScreenOn;
-		if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
-			isScreenOn = pm.isScreenOn();
-		} else {
-			isScreenOn = pm.isInteractive();
+            final boolean isScreenOn;
+            if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
+                isScreenOn = pm.isScreenOn();
+            } else {
+                isScreenOn = pm.isInteractive();
+            }
+            return isScreenOn;
+        } catch (RuntimeException e) {
+            return false;
 		}
-		return isScreenOn;
 	}
 
 	private boolean isPhoneSilenced() {
@@ -1655,8 +1659,8 @@ public class XmppConnectionService extends Service {
 	public List<Conversation> findAllConferencesWith(Contact contact) {
 		ArrayList<Conversation> results = new ArrayList<>();
 		for (final Conversation c : conversations) {
-			if (c.getMode() == Conversation.MODE_MULTI
-					&& (c.getJid().asBareJid().equals(c.getJid().asBareJid()) || c.getMucOptions().isContactInRoom(contact))) {
+			if (c.getMode() == Conversation.MODE_MULTI &&
+					(c.getJid().asBareJid().equals(contact.getJid().asBareJid()) || c.getMucOptions().isContactInRoom(contact))) {
 				results.add(c);
 			}
 		}
@@ -1802,11 +1806,7 @@ public class XmppConnectionService extends Service {
 				leaveMuc(conversation);
 			} else {
 				if (conversation.getContact().getOption(Contact.Options.PENDING_SUBSCRIPTION_REQUEST)) {
-					Log.d(Config.LOGTAG, "Canceling presence request from " + conversation.getJid().toString());
-					sendPresencePacket(
-							conversation.getAccount(),
-							mPresenceGenerator.stopPresenceUpdatesTo(conversation.getContact())
-					);
+                    stopPresenceUpdatesTo(conversation.getContact());
 				}
 			}
 			updateConversation(conversation);
@@ -1814,6 +1814,12 @@ public class XmppConnectionService extends Service {
 			updateConversationUi();
 		}
 	}
+
+    public void stopPresenceUpdatesTo(Contact contact) {
+        Log.d(Config.LOGTAG, "Canceling presence request from " + contact.getJid().toString());
+        sendPresencePacket(contact.getAccount(), mPresenceGenerator.stopPresenceUpdatesTo(contact));
+        contact.resetOption(Contact.Options.PENDING_SUBSCRIPTION_REQUEST);
+    }
 
 	private int deviceKey = -1; //ALF AM-202
 	public int getDeviceKey() {
